Rename filters->mask to clear up functionality.
